fix naming convention (server X client)  - refactored client functions to be hooks with `use` keyword in the name to clearly communicate the purpose  - remove `Server` key word from server functions
