+++ project-base/storefront/app/_components/Layout/Header/MenuIconic/MenuIconicPlaceholder.tsx
@@ -1,10 +1,10 @@
-import { getInternationalizedStaticUrlsServer } from 'app/_utils/staticUrls/getInternationalizedStaticUrlsServer';
+import { getInternationalizedStaticUrls } from 'app/_utils/getInternationalizedStaticUrls';
 import { ExtendedNextLink } from 'components/Basic/ExtendedNextLink/ExtendedNextLink';
 import { CompareIcon } from 'components/Basic/Icon/CompareIcon';
 import { HeartIcon } from 'components/Basic/Icon/HeartIcon';
 import { MarkerIcon } from 'components/Basic/Icon/MarkerIcon';
 import { UserIcon } from 'components/Basic/Icon/UserIcon';
-import useTranslation from 'next-translate/useTranslation';
+import { useTranslation } from 'components/providers/TranslationProvider';
 import Skeleton from 'react-loading-skeleton';
 import { twMergeCustom } from 'utils/twMerge';
 
@@ -13,7 +13,7 @@ const placeholderItemTwClass =
 
 export const MenuIconicPlaceholder: FC = () => {
     const { t } = useTranslation();
-    const [storesUrl] = getInternationalizedStaticUrlsServer(['/stores']);
+    const [storesUrl] = getInternationalizedStaticUrls(['/stores']);
 
     return (
         <ul className="flex lg:gap-7">
