+++ project-base/storefront/app/_components/Layout/Header/MenuIconic/SalesRepresentative.tsx
@@ -1,8 +1,8 @@
 import { MailIcon } from 'components/Basic/Icon/MailIcon';
 import { PhoneIcon } from 'components/Basic/Icon/PhoneIcon';
 import { Image } from 'components/Basic/Image/Image';
+import { useTranslation } from 'components/providers/TranslationProvider';
 import { TypeSalesRepresentative } from 'graphql/types';
-import useTranslation from 'next-translate/useTranslation';
 import { twJoin } from 'tailwind-merge';
 import { formatPhoneNumber } from 'utils/formaters/formatPhoneNumber';
