+++ project-base/storefront/app/reset-password/page.tsx
@@ -1,11 +1,11 @@
 import { ResetPasswordForm } from 'app/_components/Blocks/ResetPasswordForm/ResetPasswordForm';
 import { Breadcrumbs } from 'app/_components/Layout/Breadcrumbs/Breadcrumbs';
+import { getTranslation } from 'app/_utils/translation/getTranslation';
 import { Webline } from 'components/Layout/Webline/Webline';
 import { TypeBreadcrumbFragment } from 'graphql/requests/breadcrumbs/fragments/BreadcrumbFragment.generated';
-import { getServerT } from 'utils/getServerTranslation';
 
 export default async function ResetPasswordPage() {
-    const t = await getServerT();
+    const t = await getTranslation();
     const breadcrumbs: TypeBreadcrumbFragment[] = [{ __typename: 'Link', name: t('Reset password'), slug: '' }];
 
     return (
